+++ src/ram/trainer.py
@@ -47,6 +47,7 @@ class Trainer:
                  train_data,
                  val_data=None,
                  shuffle_each_epoch=True,
+                 patience=None,
                  replicates=1,
                  restore=False,
                  checkpoint_prefix='ckpt',
@@ -122,6 +123,8 @@ class Trainer:
                 seed=None,
                 reshuffle_each_iteration=True)
 
+        self.patience = patience
+
         # are we restoring a previous model?
         self.restore = restore
         # or are we training for a certain number of replicates?
@@ -186,6 +189,7 @@ class Trainer:
                    train_data=train_data,
                    val_data=val_data,
                    shuffle_each_epoch=config.train.shuffle_each_epoch,
+                   patience=config.train.patience,
                    replicates=config.train.replicates,
                    restore=config.train.restore,
                    checkpoint_prefix=config.train.checkpoint_prefix,
@@ -339,8 +343,13 @@ class Trainer:
                                      'without one of loss functions taking on nan values.')
 
     def _train_one_model(self):
-        """train one RAM model. Gets run once every replicate.
-        """
+        """train one RAM model. Gets run once every replicate."""
+        if self.patience:
+            max_acc = 0
+            train_accs = []
+            if self.val_data:
+                val_accs = []
+
         for epoch in range(self.epochs):
             self.logger.info(
                 f'\nEpoch: {epoch + 1}/{self.epochs} - learning rate: {self.learning_rate:.6f}'
@@ -355,7 +364,7 @@ class Trainer:
             else:
                 save_examples = False
 
-            (mn_acc_dict,
+            (acc_dict,
              losses,
              mn_loss,
              loss_went_nan) = self._train_one_epoch(current_epoch=epoch, save_examples=save_examples)
@@ -365,20 +374,41 @@ class Trainer:
             else:
                 # violating DRY by unpacking dict into vars,
                 # because apparently format strings with dict keys blow up the PyCharm parser
-                mn_acc = mn_acc_dict['mn_acc']
-                if 'mn_val_acc' in mn_acc_dict:
-                    mn_val_acc = mn_acc_dict['mn_val_acc']
-                    self.logger.info(f'mean accuracy: {mn_acc}\n'
-                                     f'mean validation accuracy: {mn_val_acc}\n'
+                train_acc = acc_dict['train']
+                train_accs.append(train_acc)
+                if 'val' in acc_dict:
+                    val_acc = acc_dict['val']
+                    val_accs.append(val_acc)
+                    self.logger.info(f'training accuracy: {train_acc: 6.3f}\n'
+                                     f'validation accuracy: {val_acc: 6.3f}\n'
                                      f'mean losses: {mn_loss}')
                 else:
-                    self.logger.info(f'mean accuracy: {mn_acc}\nmean losses: {mn_loss}')
+                    self.logger.info(f'training accuracy: {train_acc}\nmean losses: {mn_loss}')
                 self.save_checkpoint(checkpoint_path=self.data_dirs['checkpoint_path'])
+
                 if self.save_loss:
                     for loss_name, loss_arr in losses._asdict().items():
                         loss_filename = os.path.join(self.data_dirs['loss_dir'],
                                                      f'{loss_name}_epoch_{epoch}')
                         np.save(loss_filename, loss_arr)
+
+            if self.patience:
+                if 'val' in acc_dict:
+                    if acc_dict['val'] > max_acc:
+                        max_acc = acc_dict['val']
+                    if np.all(max_acc > np.asarray(val_accs[-self.patience:])):
+                        self.logger.info(f'patience is set to {self.patience}, and accuracy on validation set has not'
+                                         f' improved in {self.patience} epochs; stopping training.')
+                        self.save_checkpoint(checkpoint_path=self.data_dirs['checkpoint_path'])
+                        break
+                else:
+                    if acc_dict['train'] > max_acc:
+                        max_acc = acc_dict['val']
+                    if np.all(max_acc > np.asarray(train_accs[-self.patience:])):
+                        self.logger.info(f'patience is set to {self.patience}, and accuracy on training set has not'
+                                         f' improved in {self.patience} epochs; stopping training.')
+                        self.save_checkpoint(checkpoint_path=self.data_dirs['checkpoint_path'])
+                        break
 
         return loss_went_nan
 
@@ -620,7 +650,7 @@ class Trainer:
                 )
                 progress_bar.update(self.batch_size)
 
-        mn_acc_dict = {'mn_acc': np.mean(accs)}
+        acc_dict = {'train': np.mean(accs)}
         if not loss_is_nan:
             if save_examples:
                 for arr, stem in zip(
@@ -658,8 +688,7 @@ class Trainer:
                         val_acc = np.sum(val_acc.numpy()) / val_acc.numpy().shape[-1] * 100
                     progress_bar.update(self.batch_size)
                     val_accs.append(val_acc)
-                mn_val_acc = np.mean(val_accs)
-                mn_acc_dict['mn_val_acc'] = mn_val_acc
+                acc_dict['val'] = np.mean(val_accs)
 
         mn_loss_reinforce = np.asarray(losses_reinforce).mean()
         mn_loss_baseline = np.asarray(losses_baseline).mean()
@@ -671,4 +700,4 @@ class Trainer:
                              np.asarray(losses_action),
                              np.asarray(losses_hybrid))
         mn_loss = MeanLossTuple(mn_loss_reinforce, mn_loss_baseline, mn_loss_action, mn_losses_hybrid)
-        return mn_acc_dict, losses, mn_loss, loss_is_nan
+        return acc_dict, losses, mn_loss, loss_is_nan